dev (dot edit content): add and clean test cases
